requirements.txt
# YouTube API
google-api-python-client==2.111.0
youtube-transcript-api==1.0.0

# Data validation and models
pydantic==2.10.0
pydantic-settings==2.7.0

# Configuration
python-dotenv==1.0.1

# CLI
typer==0.9.4
rich==13.7.1

# Progress tracking
tqdm==4.66.4

# Utilities
isodate==0.6.1

# Web server
flask==3.0.3
gunicorn==22.0.0

# Audio download and transcription (optional fallback)
yt-dlp==2024.12.23
openai==1.58.1
ffmpeg-python==0.2.0

# RAG Chatbot
anthropic==0.42.0
pinecone==3.0.2

# Database (PostgreSQL for Railway)
psycopg2-binary==2.9.10

# Retry logic
tenacity==8.5.0

# Reranking
cohere==5.13.3

# PDF and web extraction
pdfplumber==0.11.4
beautifulsoup4==4.12.3
lxml==5.3.0
requests==2.32.3

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0
